@@ -1,6 +1,6 @@
 {
   "name": "@docusaurus/plugin-content-docs",
-  "version": "2.0.0-alpha.50",
+  "version": "2.0.0-alpha.56",
   "description": "Docs content plugin for Docusaurus",
   "main": "lib/index.js",
   "scripts": {
@@ -11,13 +11,13 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "@docusaurus/types": "^2.0.0-alpha.50",
+    "@docusaurus/types": "^2.0.0-alpha.56",
     "commander": "^5.0.0",
     "picomatch": "^2.1.1"
   },
   "dependencies": {
-    "@docusaurus/mdx-loader": "^2.0.0-alpha.50",
-    "@docusaurus/utils": "^2.0.0-alpha.50",
+    "@docusaurus/mdx-loader": "^2.0.0-alpha.56",
+    "@docusaurus/utils": "^2.0.0-alpha.56",
     "execa": "^3.4.0",
     "fs-extra": "^8.1.0",
     "globby": "^10.0.1",
@@ -27,7 +27,8 @@
     "lodash.groupby": "^4.6.0",
     "lodash.pick": "^4.4.0",
     "lodash.pickby": "^4.6.0",
-    "shelljs": "^0.8.3"
+    "remark-admonitions": "^1.2.1",
+    "shelljs": "^0.8.4"
   },
   "peerDependencies": {
     "@docusaurus/core": "^2.0.0",
@@ -35,6 +36,6 @@
     "react-dom": "^16.8.4"
   },
   "engines": {
-    "node": ">=10.9.0"
+    "node": ">=10.15.1"
   }
 }
