@@ -1,6 +1,6 @@
 {
   "name": "@docusaurus/plugin-content-pages",
-  "version": "2.0.0-alpha.50",
+  "version": "2.0.0-alpha.56",
   "description": "Pages content plugin for Docusaurus",
   "main": "lib/index.js",
   "scripts": {
@@ -11,8 +11,8 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@docusaurus/types": "^2.0.0-alpha.50",
-    "@docusaurus/utils": "^2.0.0-alpha.50",
+    "@docusaurus/types": "^2.0.0-alpha.56",
+    "@docusaurus/utils": "^2.0.0-alpha.56",
     "globby": "^10.0.1"
   },
   "peerDependencies": {
@@ -21,6 +21,6 @@
     "react-dom": "^16.8.4"
   },
   "engines": {
-    "node": ">=10.9.0"
+    "node": ">=10.15.1"
   }
 }
