@@ -1,6 +1,6 @@
 {
   "name": "@docusaurus/plugin-ideal-image",
-  "version": "2.0.0-alpha.50",
+  "version": "2.0.0-alpha.56",
   "description": "Docusaurus Plugin to generate an almost ideal image (responsive, lazy-loading, and low quality placeholder)",
   "main": "lib/index.js",
   "scripts": {
@@ -11,11 +11,11 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "@docusaurus/types": "^2.0.0-alpha.50",
+    "@docusaurus/types": "^2.0.0-alpha.56",
     "fs-extra": "^9.0.0"
   },
   "dependencies": {
-    "@docusaurus/lqip-loader": "^2.0.0-alpha.50",
+    "@docusaurus/lqip-loader": "^2.0.0-alpha.56",
     "@endiliey/react-ideal-image": "^0.0.11",
     "@endiliey/responsive-loader": "^1.3.2",
     "react-waypoint": "^9.0.2",
@@ -27,6 +27,6 @@
     "react-dom": "^16.8.4"
   },
   "engines": {
-    "node": ">=10.9.0"
+    "node": ">=10.15.1"
   }
 }
