@@ -63,7 +63,10 @@ module.exports = {
   },
   overrides: [
     {
-      files: ['packages/docusaurus-init/templates/classic/**/*.js'],
+      files: [
+        'packages/docusaurus-init/templates/bootstrap/**/*.js',
+        'packages/docusaurus-init/templates/classic/**/*.js',
+      ],
       rules: {
         'header/header': 'off',
       },
