@@ -5,10 +5,9 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import fs from 'fs-extra';
 import path from 'path';
 import {
-  parse,
+  parseMarkdownFile,
   aliasedSitePath,
   normalizeUrl,
   getEditUrl,
@@ -65,7 +64,7 @@ export default async function processMetadata({
   const {versioning} = env;
   const filePath = path.join(refDir, source);
 
-  const fileStringPromise = fs.readFile(filePath, 'utf-8');
+  const fileMarkdownPromise = parseMarkdownFile(filePath);
   const lastUpdatedPromise = lastUpdated(filePath, options);
 
   let version;
@@ -92,19 +91,24 @@ export default async function processMetadata({
 
   const docsEditUrl = getEditUrl(relativePath, editUrl);
 
-  const {frontMatter = {}, excerpt} = parse(await fileStringPromise);
+  const {frontMatter = {}, excerpt} = await fileMarkdownPromise;
   const {sidebar_label, custom_edit_url} = frontMatter;
 
   // Default base id is the file name.
   const baseID: string =
     frontMatter.id || path.basename(source, path.extname(source));
-
   if (baseID.includes('/')) {
     throw new Error('Document id cannot include "/".');
   }
 
-  // Append subdirectory as part of id.
+  const baseSlug: string = frontMatter.slug || baseID;
+  if (baseSlug.includes('/')) {
+    throw new Error('Document slug cannot include "/".');
+  }
+
+  // Append subdirectory as part of id/slug.
   const id = dirName !== '.' ? `${dirName}/${baseID}` : baseID;
+  const slug = dirName !== '.' ? `${dirName}/${baseSlug}` : baseSlug;
 
   // Default title is the id.
   const title: string = frontMatter.title || baseID;
@@ -114,8 +118,9 @@ export default async function processMetadata({
   // The last portion of the url path. Eg: 'foo/bar', 'bar'.
   const routePath =
     version && version !== 'next'
-      ? id.replace(new RegExp(`^version-${version}/`), '')
-      : id;
+      ? slug.replace(new RegExp(`^version-${version}/`), '')
+      : slug;
+
   const permalink = normalizeUrl([
     baseUrl,
     routeBasePath,
