@@ -25,6 +25,9 @@ function isCategoryShorthand(
   return typeof item !== 'string' && !item.type;
 }
 
+// categories are collapsed by default, unless user set collapsed = false
+const defaultCategoryCollapsedValue = true;
+
 /**
  * Convert {category1: [item1,item2]} shorthand syntax to long-form syntax
  */
@@ -33,6 +36,7 @@ function normalizeCategoryShorthand(
 ): SidebarItemCategoryRaw[] {
   return Object.entries(sidebar).map(([label, items]) => ({
     type: 'category',
+    collapsed: defaultCategoryCollapsedValue,
     label,
     items,
   }));
@@ -56,7 +60,7 @@ function assertItem(item: Object, keys: string[]): void {
 }
 
 function assertIsCategory(item: any): asserts item is SidebarItemCategoryRaw {
-  assertItem(item, ['items', 'label']);
+  assertItem(item, ['items', 'label', 'collapsed']);
   if (typeof item.label !== 'string') {
     throw new Error(
       `Error loading ${JSON.stringify(item)}. "label" must be a string.`,
@@ -65,6 +69,12 @@ function assertIsCategory(item: any): asserts item is SidebarItemCategoryRaw {
   if (!Array.isArray(item.items)) {
     throw new Error(
       `Error loading ${JSON.stringify(item)}. "items" must be an array.`,
+    );
+  }
+  // "collapsed" is an optional property
+  if (item.hasOwnProperty('collapsed') && typeof item.collapsed !== 'boolean') {
+    throw new Error(
+      `Error loading ${JSON.stringify(item)}. "collapsed" must be a boolean.`,
     );
   }
 }
@@ -111,7 +121,13 @@ function normalizeItem(item: SidebarItemRaw): SidebarItem[] {
   switch (item.type) {
     case 'category':
       assertIsCategory(item);
-      return [{...item, items: flatMap(item.items, normalizeItem)}];
+      return [
+        {
+          collapsed: defaultCategoryCollapsedValue,
+          ...item,
+          items: flatMap(item.items, normalizeItem),
+        },
+      ];
     case 'link':
       assertIsLink(item);
       return [item];
@@ -120,7 +136,15 @@ function normalizeItem(item: SidebarItemRaw): SidebarItem[] {
       assertIsDoc(item);
       return [item];
     default:
-      throw new Error(`Unknown sidebar item type: ${item.type}`);
+      const extraMigrationError =
+        item.type === 'subcategory'
+          ? "Docusaurus v2: 'subcategory' has been renamed as 'category'"
+          : '';
+      throw new Error(
+        `Unknown sidebar item type [${
+          item.type
+        }]. Sidebar item=${JSON.stringify(item)} ${extraMigrationError}`,
+      );
   }
 }
 
