@@ -21,8 +21,10 @@ export interface PluginOptions extends MetadataOptions, PathOptions {
   include: string[];
   docLayoutComponent: string;
   docItemComponent: string;
-  remarkPlugins: string[];
+  remarkPlugins: ([Function, object] | Function)[];
   rehypePlugins: string[];
+  admonitions: any;
+  homePageId: string;
 }
 
 export type SidebarItemDoc = {
@@ -40,12 +42,14 @@ export interface SidebarItemCategory {
   type: 'category';
   label: string;
   items: SidebarItem[];
+  collapsed: boolean;
 }
 
 export interface SidebarItemCategoryRaw {
   type: 'category';
   label: string;
   items: SidebarItemRaw[];
+  collapsed?: boolean;
 }
 
 export type SidebarItem =
@@ -81,6 +85,7 @@ export interface DocsSidebarItemCategory {
   type: 'category';
   label: string;
   items: DocsSidebarItem[];
+  collapsed?: boolean;
 }
 
 export type DocsSidebarItem = SidebarItemLink | DocsSidebarItemCategory;
@@ -159,6 +164,8 @@ export type DocsBaseMetadata = Pick<
   'docsSidebars' | 'permalinkToSidebar'
 > & {
   version?: string;
+  isHomePage?: boolean;
+  homePagePath?: string;
 };
 
 export type VersioningEnv = {
