@@ -30,7 +30,22 @@ describe('createSitemap', () => {
     expect(() => {
       createSitemap({} as any, [], {} as any);
     }).toThrowErrorMatchingInlineSnapshot(
-      `"Url in docusaurus.config.js cannot be empty/undefined"`,
+      `"url in docusaurus.config.js cannot be empty/undefined"`,
     );
+  });
+
+  test('exclusion of 404 page', () => {
+    const sitemap = createSitemap(
+      {
+        url: 'https://example.com',
+      } as DocusaurusConfig,
+      ['/', '/404.html', '/mypage'],
+      {
+        cacheTime: 600,
+        changefreq: 'daily',
+        priority: 0.7,
+      },
+    );
+    expect(sitemap.toString()).not.toContain('404');
   });
 });
