@@ -6,21 +6,6 @@
  */
 
 import Prism from 'prism-react-renderer/prism';
-import ExecutionEnvironment from '@docusaurus/ExecutionEnvironment';
-import siteConfig from '@generated/docusaurus.config';
+import prismIncludeLanguages from '@theme/prism-include-languages';
 
-(() => {
-  if (ExecutionEnvironment.canUseDOM) {
-    const {
-      themeConfig: {prism: {additionalLanguages = []} = {}},
-    } = siteConfig;
-
-    window.Prism = Prism;
-
-    additionalLanguages.forEach((lang) => {
-      require(`prismjs/components/prism-${lang}`); // eslint-disable-line
-    });
-
-    delete window.Prism;
-  }
-})();
+prismIncludeLanguages(Prism);
