@@ -49,12 +49,12 @@ function BlogTagsListPage(props) {
 
   return (
     <Layout title="Tags" description="Blog Tags">
-      <div className="container margin-vert--xl">
+      <div className="container margin-vert--lg">
         <div className="row">
-          <div className="col col--8 col--offset-2">
+          <main className="col col--8 col--offset-2">
             <h1>Tags</h1>
             <div className="margin-vert--lg">{tagsSection}</div>
-          </div>
+          </main>
         </div>
       </div>
     </Layout>
