@@ -23,9 +23,9 @@ function BlogTagsPostPage(props) {
     <Layout
       title={`Posts tagged "${tagName}"`}
       description={`Blog | Tagged "${tagName}"`}>
-      <div className="container margin-vert--xl">
+      <div className="container margin-vert--lg">
         <div className="row">
-          <div className="col col--8 col--offset-2">
+          <main className="col col--8 col--offset-2">
             <h1>
               {count} {pluralize(count, 'post')} tagged with &quot;{tagName}
               &quot;
@@ -42,7 +42,7 @@ function BlogTagsPostPage(props) {
                 </BlogPostItem>
               ))}
             </div>
-          </div>
+          </main>
         </div>
       </div>
     </Layout>
