@@ -8,13 +8,12 @@
 /* eslint-disable jsx-a11y/no-noninteractive-tabindex */
 
 import React, {useEffect, useState, useRef} from 'react';
-import classnames from 'classnames';
+import clsx from 'clsx';
 import Highlight, {defaultProps} from 'prism-react-renderer';
-import defaultTheme from 'prism-react-renderer/themes/palenight';
-import Clipboard from 'clipboard';
+import copy from 'copy-text-to-clipboard';
 import rangeParser from 'parse-numeric-range';
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
-import useThemeContext from '@theme/hooks/useThemeContext';
+import usePrismTheme from '@theme/hooks/usePrismTheme';
 
 import styles from './styles.module.css';
 
@@ -108,15 +107,11 @@ export default ({children, className: languageClassName, metastring}) => {
     setMounted(true);
   }, []);
 
-  const target = useRef(null);
   const button = useRef(null);
   let highlightLines = [];
   let codeBlockTitle = '';
 
-  const {isDarkTheme} = useThemeContext();
-  const lightModeTheme = prism.theme || defaultTheme;
-  const darkModeTheme = prism.darkTheme || lightModeTheme;
-  const prismTheme = isDarkTheme ? darkModeTheme : lightModeTheme;
+  const prismTheme = usePrismTheme();
 
   if (metastring && highlightLinesRangeRegex.test(metastring)) {
     const highlightLinesRange = metastring.match(highlightLinesRangeRegex)[1];
@@ -131,22 +126,6 @@ export default ({children, className: languageClassName, metastring}) => {
       .split('title=')[1]
       .replace(/"+/g, '');
   }
-
-  useEffect(() => {
-    let clipboard;
-
-    if (button.current) {
-      clipboard = new Clipboard(button.current, {
-        target: () => target.current,
-      });
-    }
-
-    return () => {
-      if (clipboard) {
-        clipboard.destroy();
-      }
-    };
-  }, [button.current, target.current]);
 
   let language =
     languageClassName && languageClassName.replace(/language-/, '');
@@ -200,7 +179,7 @@ export default ({children, className: languageClassName, metastring}) => {
   }
 
   const handleCopyCode = () => {
-    window.getSelection().empty();
+    copy(code);
     setShowCopied(true);
 
     setTimeout(() => setShowCopied(false), 2000);
@@ -225,7 +204,7 @@ export default ({children, className: languageClassName, metastring}) => {
               ref={button}
               type="button"
               aria-label="Copy code to clipboard"
-              className={classnames(styles.copyButton, {
+              className={clsx(styles.copyButton, {
                 [styles.copyButtonWithTitle]: codeBlockTitle,
               })}
               onClick={handleCopyCode}>
@@ -233,10 +212,10 @@ export default ({children, className: languageClassName, metastring}) => {
             </button>
             <div
               tabIndex="0"
-              className={classnames(className, styles.codeBlock, {
+              className={clsx(className, styles.codeBlock, {
                 [styles.codeBlockWithTitle]: codeBlockTitle,
               })}>
-              <div ref={target} className={styles.codeBlockLines} style={style}>
+              <div className={styles.codeBlockLines} style={style}>
                 {tokens.map((line, i) => {
                   if (line.length === 1 && line[0].content === '') {
                     line[0].content = '\n'; // eslint-disable-line no-param-reassign
