@@ -7,12 +7,11 @@
 
 import React from 'react';
 import Head from '@docusaurus/Head';
-import isInternalUrl from '@docusaurus/isInternalUrl';
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
 import useBaseUrl from '@docusaurus/useBaseUrl';
 
 import ThemeProvider from '@theme/ThemeProvider';
-import TabGroupChoiceProvider from '@theme/TabGroupChoiceProvider';
+import UserPreferencesProvider from '@theme/UserPreferencesProvider';
 import AnnouncementBar from '@theme/AnnouncementBar';
 import Navbar from '@theme/Navbar';
 import Footer from '@theme/Footer';
@@ -38,18 +37,13 @@ function Layout(props) {
     version,
   } = props;
   const metaTitle = title ? `${title} | ${siteTitle}` : siteTitle;
-
   const metaImage = image || defaultImage;
-  let metaImageUrl = siteUrl + useBaseUrl(metaImage);
-  if (!isInternalUrl(metaImage)) {
-    metaImageUrl = metaImage;
-  }
-
+  const metaImageUrl = useBaseUrl(metaImage, {absolute: true});
   const faviconUrl = useBaseUrl(favicon);
 
   return (
     <ThemeProvider>
-      <TabGroupChoiceProvider>
+      <UserPreferencesProvider>
         <Head>
           {/* TODO: Do not assume that it is in english language */}
           <html lang="en" />
@@ -75,13 +69,14 @@ function Layout(props) {
           {permalink && (
             <meta property="og:url" content={siteUrl + permalink} />
           )}
+          {permalink && <link rel="canonical" href={siteUrl + permalink} />}
           <meta name="twitter:card" content="summary_large_image" />
         </Head>
         <AnnouncementBar />
         <Navbar />
         <div className="main-wrapper">{children}</div>
         {!noFooter && <Footer />}
-      </TabGroupChoiceProvider>
+      </UserPreferencesProvider>
     </ThemeProvider>
   );
 }
