@@ -10,11 +10,11 @@ import Toggle from 'react-toggle';
 
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
 
-import classnames from 'classnames';
+import clsx from 'clsx';
 import styles from './styles.module.css';
 
-const Moon = () => <span className={classnames(styles.toggle, styles.moon)} />;
-const Sun = () => <span className={classnames(styles.toggle, styles.sun)} />;
+const Moon = () => <span className={clsx(styles.toggle, styles.moon)} />;
+const Sun = () => <span className={clsx(styles.toggle, styles.sun)} />;
 
 export default function (props) {
   const {isClient} = useDocusaurusContext();
