@@ -8,6 +8,7 @@
 import {useState, useCallback, useEffect} from 'react';
 import {useLocation} from '@docusaurus/router';
 import useLocationHash from '@theme/hooks/useLocationHash';
+import useScrollPosition from '@theme/hooks/useScrollPosition';
 
 const useHideableNavbar = (hideOnScroll) => {
   const [isNavbarVisible, setIsNavbarVisible] = useState(true);
@@ -22,47 +23,41 @@ const useHideableNavbar = (hideOnScroll) => {
   const location = useLocation();
   const [hash, setHash] = useLocationHash(location.hash);
 
-  const handleScroll = () => {
-    const scrollTop = window.pageYOffset || document.documentElement.scrollTop;
+  useScrollPosition(
+    ({scrollY: scrollTop}) => {
+      if (!hideOnScroll) {
+        return;
+      }
 
-    if (scrollTop === 0) {
-      setIsNavbarVisible(true);
-    }
+      if (scrollTop === 0) {
+        setIsNavbarVisible(true);
+      }
 
-    if (scrollTop < navbarHeight) {
-      return;
-    }
+      if (scrollTop < navbarHeight) {
+        return;
+      }
 
-    if (isFocusedAnchor) {
-      setIsFocusedAnchor(false);
-      setIsNavbarVisible(false);
+      if (isFocusedAnchor) {
+        setIsFocusedAnchor(false);
+        setIsNavbarVisible(false);
+        setLastScrollTop(scrollTop);
+        return;
+      }
+
+      const documentHeight =
+        document.documentElement.scrollHeight - navbarHeight;
+      const windowHeight = window.innerHeight;
+
+      if (lastScrollTop && scrollTop >= lastScrollTop) {
+        setIsNavbarVisible(false);
+      } else if (scrollTop + windowHeight < documentHeight) {
+        setIsNavbarVisible(true);
+      }
+
       setLastScrollTop(scrollTop);
-      return;
-    }
-
-    const documentHeight = document.documentElement.scrollHeight - navbarHeight;
-    const windowHeight = window.innerHeight;
-
-    if (lastScrollTop && scrollTop >= lastScrollTop) {
-      setIsNavbarVisible(false);
-    } else if (scrollTop + windowHeight < documentHeight) {
-      setIsNavbarVisible(true);
-    }
-
-    setLastScrollTop(scrollTop);
-  };
-
-  useEffect(() => {
-    if (!hideOnScroll) {
-      return undefined;
-    }
-
-    window.addEventListener('scroll', handleScroll);
-
-    return () => {
-      window.removeEventListener('scroll', handleScroll);
-    };
-  }, [lastScrollTop, navbarHeight]);
+    },
+    [lastScrollTop, navbarHeight],
+  );
 
   useEffect(() => {
     if (!hideOnScroll) {
