@@ -7,27 +7,27 @@
 
 import * as React from 'react';
 import {LiveProvider, LiveEditor, LiveError, LivePreview} from 'react-live';
-import classnames from 'classnames';
+import clsx from 'clsx';
 
 import styles from './styles.module.css';
 
 function Playground({children, theme, transformCode, ...props}) {
   return (
     <LiveProvider
-      code={children}
+      code={children.replace(/\n$/, '')}
       transformCode={transformCode || ((code) => `${code};`)}
       theme={theme}
       {...props}>
       <div
-        className={classnames(
+        className={clsx(
           styles.playgroundHeader,
           styles.playgroundEditorHeader,
         )}>
         Live Editor
       </div>
-      <LiveEditor />
+      <LiveEditor className={styles.playgroundEditor} />
       <div
-        className={classnames(
+        className={clsx(
           styles.playgroundHeader,
           styles.playgroundPreviewHeader,
         )}>
