@@ -6,13 +6,29 @@
  */
 
 const path = require('path');
+const fs = require('fs');
+const eta = require('eta');
+const {normalizeUrl} = require('@docusaurus/utils');
+const openSearchTemplate = require('./templates/opensearch');
 
-module.exports = function () {
+const OPEN_SEARCH_FILENAME = 'opensearch.xml';
+
+module.exports = function (context) {
+  const {
+    baseUrl,
+    siteConfig: {title, url, favicon},
+  } = context;
+  const pagePath = path.resolve(__dirname, './pages/search/index.js');
+
   return {
     name: 'docusaurus-theme-search-algolia',
 
     getThemePath() {
       return path.resolve(__dirname, './theme');
+    },
+
+    getPathsToWatch() {
+      return [pagePath];
     },
 
     configureWebpack() {
@@ -32,6 +48,45 @@ module.exports = function () {
             },
           },
         },
+      };
+    },
+
+    async contentLoaded({actions: {addRoute}}) {
+      addRoute({
+        path: normalizeUrl([baseUrl, 'search']),
+        component: pagePath,
+        exact: true,
+      });
+    },
+
+    async postBuild({outDir}) {
+      try {
+        fs.writeFileSync(
+          path.join(outDir, OPEN_SEARCH_FILENAME),
+          eta.render(openSearchTemplate.trim(), {
+            title,
+            url,
+            favicon: normalizeUrl([url, favicon]),
+          }),
+        );
+      } catch (err) {
+        throw new Error(`Generating OpenSearch file failed: ${err}`);
+      }
+    },
+
+    injectHtmlTags() {
+      return {
+        headTags: [
+          {
+            tagName: 'link',
+            attributes: {
+              rel: 'search',
+              type: 'application/opensearchdescription+xml',
+              title,
+              href: normalizeUrl([baseUrl, OPEN_SEARCH_FILENAME]),
+            },
+          },
+        ],
       };
     },
   };
