@@ -15,6 +15,6 @@ const srcDir = path.resolve(__dirname, 'src');
 const libDir = path.resolve(__dirname, 'lib');
 fs.copySync(srcDir, libDir, {
   filter(filepath) {
-    return !/__tests__/.test(filepath) && !/\.ts$/.test(filepath);
+    return !/__tests__/.test(filepath) && !/\.tsx?$/.test(filepath);
   },
 });
