@@ -5,8 +5,11 @@
  * LICENSE file in the root directory of this source tree.
  */
 
+import Module from 'module';
+import {join} from 'path';
 import importFresh from 'import-fresh';
 import {LoadContext, Plugin, PluginConfig} from '@docusaurus/types';
+import {CONFIG_FILE_NAME} from '../../constants';
 
 export function initPlugins({
   pluginConfigs,
@@ -15,9 +18,16 @@ export function initPlugins({
   pluginConfigs: PluginConfig[];
   context: LoadContext;
 }): Plugin<any>[] {
+  // We need to resolve plugins from the perspective of the siteDir, since the siteDir's package.json
+  // declares the dependency on these plugins.
+  // We need to fallback to createRequireFromPath since createRequire is only available in node v12.
+  // See: https://nodejs.org/api/modules.html#modules_module_createrequire_filename
+  const createRequire = Module.createRequire || Module.createRequireFromPath;
+  const pluginRequire = createRequire(join(context.siteDir, CONFIG_FILE_NAME));
+
   const plugins: Plugin<any>[] = pluginConfigs
     .map((pluginItem) => {
-      let pluginModuleImport;
+      let pluginModuleImport: string | undefined;
       let pluginOptions = {};
 
       if (!pluginItem) {
@@ -37,7 +47,9 @@ export function initPlugins({
 
       // The pluginModuleImport value is any valid
       // module identifier - npm package or locally-resolved path.
-      const pluginModule: any = importFresh(pluginModuleImport);
+      const pluginModule: any = importFresh(
+        pluginRequire.resolve(pluginModuleImport),
+      );
       return (pluginModule.default || pluginModule)(context, pluginOptions);
     })
     .filter(Boolean);
