@@ -6,30 +6,51 @@
  */
 
 module.exports = {
-  docs: {
-    Docusaurus: ['introduction', 'design-principles', 'contributing'],
-    'Getting Started': ['installation', 'configuration'],
-    Guides: [
-      'creating-pages',
-      'styling-layout',
-      'static-assets',
-      {
-        Docs: ['docs', 'markdown-features', 'versioning'],
-      },
-      'blog',
-      'search',
-      'deployment',
-      'migrating-from-v1-to-v2',
-    ],
-    'Advanced Guides': ['using-plugins', 'using-themes', 'presets'],
-    'API Reference': [
-      'cli',
-      'docusaurus-core',
-      'docusaurus.config.js',
-      'lifecycle-apis',
-      'theme-classic',
-    ],
-  },
+  docs: [
+    {
+      type: 'category',
+      label: 'Docusaurus',
+      items: ['introduction', 'design-principles', 'contributing'],
+    },
+    {
+      type: 'category',
+      label: 'Getting Started',
+      collapsed: false,
+      items: ['installation', 'configuration'],
+    },
+    {
+      type: 'category',
+      label: 'Guides',
+      items: [
+        'creating-pages',
+        'styling-layout',
+        'static-assets',
+        {
+          Docs: ['docs-introduction', 'markdown-features', 'versioning'],
+        },
+        'blog',
+        'search',
+        'deployment',
+        'migrating-from-v1-to-v2',
+      ],
+    },
+    {
+      type: 'category',
+      label: 'Advanced Guides',
+      items: ['using-plugins', 'using-themes', 'presets'],
+    },
+    {
+      type: 'category',
+      label: 'API Reference',
+      items: [
+        'cli',
+        'docusaurus-core',
+        'docusaurus.config.js',
+        'lifecycle-apis',
+        'theme-classic',
+      ],
+    },
+  ],
   community: [
     'support',
     'team',
