@@ -9,7 +9,7 @@ import React, {useEffect} from 'react';
 import Layout from '@theme/Layout';
 
 import canny from '../../scripts/canny';
-import classnames from 'classnames';
+import clsx from 'clsx';
 import styles from './styles.module.css';
 
 const BOARD_TOKEN = '054e0e53-d951-b14c-7e74-9eb8f9ed2f91';
@@ -29,8 +29,8 @@ function Feedback() {
       permalink="/feedback"
       title="Feedback"
       description="Docusaurus 2 Feedback page">
-      <div
-        className={classnames(
+      <main
+        className={clsx(
           'container',
           'margin-vert--xl',
           styles.feedbackBackground,
